feat: major refactor - auth flow, home screen, new catalog/profile features, and theme updates

+++ app/src/main/java/com/lecomapp/liquefied/features/auth/data/datasources/remote/dto/AuthResponse.kt
@@ -26,7 +26,7 @@ data class UserResponse(
     val city: String? = null,
     val state: String? = null,
     val country: String? = null,
-    val zipCode: Int? = null,
+    val zipCode: Long? = null,
     val isActive: Boolean,
     val isEmailVerified: Boolean,
     val isPhoneVerified: Boolean,
